% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum.R
\name{C2beta}
\alias{C2beta}
\title{Compute the coefficients from the continuum regression (CR) algorithm}
\usage{
C2beta(X, Y, C, lambda)
}
\arguments{
\item{X}{The input feature matrix}

\item{Y}{The input response vector}

\item{C}{The weight matrix computed from CR algorithm}

\item{lambda}{Deprecated. Regularization parameter if L2 penalization is used for CR.
JICO uses zero as default.}
}
\value{
A list of regression coefficients to perform the prediction task.
}
\description{
This function converts the CR algorithm outputs to the regression coefficients
}
