% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createFolds}
\alias{createFolds}
\title{Utility function to create folds for stratified samples}
\usage{
createFolds(strat_id, k)
}
\arguments{
\item{strat_id}{A vector of the stratified sample id.
E.g. In total of 5 samples, first three from group 1, last two from group 2
-> c(1, 1, 1, 2, 2)}

\item{k}{Number of folds to create.}
}
\value{
A list of sample indices in k folds.
}
\description{
This function generate data folds for cross validation given stratified samples
}
