% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{DIAG}
\alias{DIAG}
\title{Generate diagonal matrix}
\usage{
DIAG(e)
}
\arguments{
\item{e}{Diagonal element. Can be a vector or a number}
}
\value{
A square diagonal matrix using the input as diagonal elements
}
\description{
This function returns a diagnoal matrix using the input vector or number as diagonal.
}
