% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuum.R
\name{initialize.UDVZ}
\alias{initialize.UDVZ}
\title{Helper function to compute the SVD results}
\usage{
initialize.UDVZ(X)
}
\arguments{
\item{X}{The input feature matrix}
}
\value{
A list of SVD results that are served as CR algorithm's inputs.
}
\description{
This function computes the SVD results from a given matrix X.
This is used as the initialization for the continuum regression.
}
