% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter.set.G_3}
\alias{parameter.set.G_3}
\title{Generate parameter sets (G=3)}
\usage{
parameter.set.G_3(maxrankA, maxrankJ, gamma)
}
\arguments{
\item{maxrankA}{The maximum rank for individual component}

\item{maxrankJ}{The maximum rank for joint component}

\item{gamma}{The gamma parameter. Need to be fixed.}
}
\value{
A list of hyperparameter candidates
}
\description{
This function generate set of hyperparameters when there
are three groups.
}
