% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parameter.set.rankA_eq}
\alias{parameter.set.rankA_eq}
\title{Generate parameter sets (equal individual ranks)}
\usage{
parameter.set.rankA_eq(G, maxrankA, maxrankJ, gamma.list)
}
\arguments{
\item{G}{number of groups}

\item{maxrankA}{The maximum rank for individual component}

\item{maxrankJ}{The maximum rank for joint component}

\item{gamma.list}{The list of candidate gammas to be tuned}
}
\value{
A list of hyperparameter candidates
}
\description{
This function generate set of hyperparameters when the individual ranks are the same
}
