% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SOLVE}
\alias{SOLVE}
\title{Helper function to compute the inverse of input X matrix}
\usage{
SOLVE(x)
}
\arguments{
\item{x}{The input matrix X}
}
\value{
Either the general inverse of X or the X itself
}
\description{
This function computes the general inverse of X when it exists.
If X contains a degenerated dimension, return the original X.
}
